+++ lib/src/category_icons.dart
