+++ lib/src/default_emoji_picker_view.dart
@@ -1,7 +1,7 @@
 import 'package:emoji_picker_flutter/emoji_picker_flutter.dart';
 import 'package:emoji_picker_flutter/src/category_emoji.dart';
-import 'package:emoji_picker_flutter/config.dart';
-import 'package:emoji_picker_flutter/emoji_picker_builder.dart';
+import 'package:emoji_picker_flutter/src/config.dart';
+import 'package:emoji_picker_flutter/src/emoji_picker_builder.dart';
 import 'package:emoji_picker_flutter/src/emoji_view_state.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
